Refaktoroi suunnitteluvaiheen luonnoksen ja aineistot. Huom! Padding/margin-hommat meni huonosti.
